+++ packages/api-explorer/src/components/Header/Header.spec.tsx
@@ -28,46 +28,35 @@ import { act, screen, waitFor } from '@testing-library/react'
 import { codeGenerators } from '@looker/sdk-codegen'
 import userEvent from '@testing-library/user-event'
 
-import { specs, specState } from '../../test-data'
+import { getLoadedSpecs, specs } from '../../test-data'
 import { renderWithRouterAndReduxProvider } from '../../test-utils'
 import { defaultSettingsState } from '../../state'
 import { Header } from './Header'
 
 describe('Header', () => {
-  const specDispatch = jest.fn()
   const hasNavigation = true
   const toggleNavigation = () => !hasNavigation
-
+  const spec = getLoadedSpecs()['4.0']
   beforeAll(() => {
     window.HTMLElement.prototype.scrollIntoView = jest.fn()
   })
 
   test('it renders a title', () => {
     renderWithRouterAndReduxProvider(
-      <Header
-        specs={specs}
-        spec={specState.spec}
-        specDispatch={specDispatch}
-        toggleNavigation={toggleNavigation}
-      />
+      <Header spec={spec} toggleNavigation={toggleNavigation} />
     )
     expect(screen.getByText('API Explorer').closest('a')).toHaveAttribute(
       'href',
-      `/${specState.spec.key}`
+      `/${spec.key}`
     )
   })
 
   test('it renders a spec selector with the correct default value and options', async () => {
     renderWithRouterAndReduxProvider(
-      <Header
-        specs={specs}
-        spec={specState.spec}
-        specDispatch={specDispatch}
-        toggleNavigation={toggleNavigation}
-      />
+      <Header spec={spec} toggleNavigation={toggleNavigation} />
     )
     const selector = screen.getByLabelText('spec selector')
-    expect(selector).toHaveValue(`${specState.spec.key}`)
+    expect(selector).toHaveValue(`${spec.key}`)
     await act(async () => {
       await userEvent.click(selector)
       await waitFor(() => {
@@ -80,12 +69,7 @@ describe('Header', () => {
 
   test('it renders an sdk language selector with the correct value and options', async () => {
     renderWithRouterAndReduxProvider(
-      <Header
-        specs={specs}
-        spec={specState.spec}
-        specDispatch={specDispatch}
-        toggleNavigation={toggleNavigation}
-      />
+      <Header spec={spec} toggleNavigation={toggleNavigation} />
     )
     const selector = screen.getByLabelText('sdk language selector')
     expect(selector).toHaveValue(defaultSettingsState.sdkLanguage)
@@ -101,12 +85,7 @@ describe('Header', () => {
 
   test('it renders an icon button for the differ', () => {
     renderWithRouterAndReduxProvider(
-      <Header
-        specs={specs}
-        spec={specState.spec}
-        specDispatch={specDispatch}
-        toggleNavigation={toggleNavigation}
-      />
+      <Header spec={spec} toggleNavigation={toggleNavigation} />
     )
     expect(
       screen
@@ -114,6 +93,6 @@ describe('Header', () => {
           name: 'Compare Specifications',
         })
         .closest('a')
-    ).toHaveAttribute('href', `/diff/${specState.spec.key}/`)
+    ).toHaveAttribute('href', `/diff/${spec.key}/`)
   })
 })