+++ packages/hackathon/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [22.0.0](https://www.github.com/looker-open-source/sdk-codegen/compare/hackathon-v21.20.1...hackathon-v22.0.0) (2022-02-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/code-editor bumped from ^0.1.17 to ^0.1.18
+    * @looker/extension-sdk-react bumped from ^21.20.1 to ^22.0.0
+    * @looker/extension-utils bumped from ^0.1.3 to ^0.1.4
+    * @looker/wholly-sheet bumped from ^0.5.23 to ^0.5.24
+
 ### [21.20.1](https://www.github.com/looker-open-source/sdk-codegen/compare/hackathon-v21.20.0...hackathon-v21.20.1) (2022-01-27)